Deployable Sencha Touch kitchensink for Qt
"Standalone" version of the kitchensink, with a bit of boiler-
plate code that sets a QWebView up and loads the kitchensink.

This project was created using the Qt Creator Mobile App Wizard
with it's deployment helper.
Runs well in Qt Creator on Windows and Symbian. Should also
work with Maemo and Linux Desktop.

This is the first version for Qt 4.6.x+.